backend/requirements.txt
Flask==3.0.0
scikit-learn==1.3.0
pandas==2.0.3
nltk==3.8.1
gunicorn==21.2.0
pillow==10.2.0
pytesseract==0.3.10
